Update SCM URLs
I was having some trouble running the PCT for this plugin, and I think part of the issue was that the with URLs as they were configured before the PCT was executing ` git clone ssh://github.com/jenkinsci/ownership-plugin.git` which I don't think can work. I'm not 100% sure how these URLs work so I might be mistaken.